Fixes an issue where the _addExtremePeriods method would fail due to MultiIndexes
When a MultiIndex is passed as columns, a tuple is as the column name for ExtremePeriods. This raises: "Can only concatenate tuple (not "str") to tuple error." since a str cannot be appended to a tuple. This fix fill append the extreme-period identifier to the last element of the tuple.